requirements.txt
altair==5.2.0
annotated-types==0.6.0
attrs==23.2.0
azure-cognitiveservices-speech==1.35.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
filelock==3.13.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
huggingface-hub==0.21.4
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
preshed==3.0.9
protobuf==4.25.3
pyarrow==15.0.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyodbc==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
sentencepiece==0.2.0
setuptools==69.1.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.31.1
streamlit-audiorecorder==0.0.4
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
wasabi==1.1.2
weasel==0.3.4
zipp==3.17.0
